Completed Projects/old streamlit camelaci frontend/requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
camel-ai>=0.1.0
rich>=13.0.0
nest-asyncio>=1.5.0 